+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## ChangeLog
 
+* Added support for custom error messages and stack traces.  Errors provided
+  to `Transaction.NoticeError` will now be checked to see if
+  they implement [ErrorClasser](https://godoc.org/github.com/newrelic/go-agent#ErrorClasser)
+  and/or [StackTracer](https://godoc.org/github.com/newrelic/go-agent#StackTracer).
+  Thanks to @fgrosse for this proposal.
+
+* Added support for [pkg/errors](https://github.com/pkg/errors).  Thanks to
+  @fgrosse for this work.
+  * [documentation](https://godoc.org/github.com/newrelic/go-agent/_integrations/nrpkgerrors)
+  * [example](https://github.com/newrelic/go-agent/blob/master/_integrations/nrpkgerrors/nrpkgerrors.go)
+
 * Fixed tests for Go 1.8.
 
 ## 1.5.0